demo/requirements.txt
numpy
cupy-cuda12x # edit to use correct CUDA version
qtpy
matplotlib
rainbow-logging-handler
